// src/scripts/enum.ts
export enum EScoreMessage {
  VERY_LOW_GRADE = 'Sua nota foi muito baixa! Estude mais um pouco.',
  LOW_GRADE = 'Sua nota foi boa, mas pode melhorar.',
  GOOD_GRADE = 'Sua nota foi excelente! Continue estudando.',
  PERFECT_GRADE = 'Parabéns! Você acertou todas as questões!',
}

export enum EQuizThemeName {
  JAVASCRIPT = 'JavaScript',
  HTML = 'HTML',
  CSS = 'CSS',
  REACT_JS = 'ReactJS',
  ANGULAR = 'Angular',
  NODE_JS = 'NodeJS',
  FRONT_END = 'FrontEnd',
  JAVA = 'Java',
  PYTHON = 'Python',
  BACK_END = 'BackEnd',
}